server/requirements.txt
alembic==1.6.5
deepl==1.13.0
Flask==2.0.1
Flask_Cors==3.0.10
Flask_JWT_Extended==4.2.3
Flask_Migrate==3.1.0
Flask_SQLAlchemy==2.5.1
matplotlib==3.5.1
nltk==3.6.7
pandas==1.3.4
python-dotenv==0.21.1
seaborn==0.11.2
spacy==3.4.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
